feat: move authentication logic to authUtils and clean up App component
